+++ OUDS/Core/Themes/Wireframe/Sources/TypeAliases/WireframeBrandColorRawTokens+Aliases.swift
@@ -1,0 +1,20 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDSTokensRaw
+
+/// In the global design system tool, the verbs of "brand colors" are used, and at this level mean simply raw tokens of colors.
+public typealias ColorWireframeBrandRawToken = ColorRawToken
+
+/// In the global design system tool, the verbs of "decorative colors" are used, and at this level mean simply raw tokens of colors, but decorative ones
+public typealias ColorWireframeDecorativeRawToken = ColorRawToken