+++ OUDS/Core/Themes/Wireframe/Tests/Values/SemanticTokens/WireframeTheme+BorderSemanticTokensTests.swift
@@ -12,21 +12,21 @@
 //
 
 import OUDS
-import OUDSThemesOrangeInverse
+import OUDSThemesWireframe
 import Testing
 
 // swiftlint:disable type_name
 
-/// The aim of this tests class is to look for regressions in **OrangeInverse brand border semantic tokens**.
+/// The aim of this tests class is to look for regressions in **Wireframe brand border semantic tokens**.
 /// Indeed, each future generation of Swift code may break theses tests because there are new values.
 /// In the semantics of **border semantic tokens**, there will be some unchanged things like relationships between tokens.
 /// Thus this tests class just checks if such relationships are still here whatever the values at the end.
-struct OrangeInverseThemeBorderSemanticTokensTests {
+struct WireframeThemeBorderSemanticTokensTests {
 
     private var theme: OUDSTheme
 
     init() {
-        theme = OrangeInverseTheme()
+        theme = WireframeTheme()
     }
 
     @Test func borderWidthSemanticTokensAreBiggerAndBigger() throws {